standup/status/migrations/0015_team_users.py
# -*- coding: utf-8 -*-
from __future__ import unicode_literals

from django.db import migrations, models


class Migration(migrations.Migration):

    dependencies = [
        ('status', '0014_remove_teamuser'),
    ]

    operations = [
        migrations.AddField(
            model_name='team',
            name='users',
            field=models.ManyToManyField(related_name='teams', to='status.StandupUser'),
        ),
    ]
