requirements-dev.txt
-r requirements.txt

pytest==3.0.1 \
    --hash=sha256:ba6ae459222c3ee650d9e7a3f8cbf69c23c20ed6977496e9081f3b8da19107d0 \
    --hash=sha256:e82bc0596ee96b2287c08705cfcb6898db1fe4b5c87db3b6823f1fdd77fb3ff1
pytest-django==3.0.0 \
    --hash=sha256:7898731205fa245926c800273dcffcd9b25e0b660a363ecb6cc2901ffe4aa1cc \
    --hash=sha256:5952283b0e6e0f9d8cb46803636e3940e1c77da8a529f03a8a4320f52d9bf560
mock==2.0.0 --hash=sha256:5ce3c71c5545b472da17b72268978914d0252980348636840bd34a00b5cc96c1
factory_boy==2.7.0 --hash=sha256:36c949d5c7adefb02d25323b7a5a97dc698e58ef84c4654845ecb2e34bee9a23
# Note: fake_factory gets changed to faker sometime after September 15th, 2016. That probably
# means we need to update fake_factory and factory_boy packages.
fake_factory==0.5.10 --hash=sha256:a3491f0a745a8bdd8199d1ffa04aeee3e38146a4cc3e7789ae785d159a9f69c4
python-dateutil==2.5.3 --hash=sha256:598499a75be2e5e18a66f12c00dd47a069de24794effeda4228bfc760f44f527
Sphinx==1.4.5 --hash=sha256:a1901af3d437473662daca2ee8f03c232d84af08a0b119d6bc55020661b12ab8
pbr==1.10.0 --hash=sha256:f5cf7265a80636ecff66806d13494cbf9d77a3758a65fd8b4d4d4bee81b0c375
py==1.4.31 --hash=sha256:4a3e4f3000c123835ac39cab5ccc510642153bc47bc1f13e2bbb53039540ae69
imagesize==0.7.1 --hash=sha256:6ebdc9e0ad188f9d1b2cdd9bc59cbe42bf931875e829e7a595e6b3abdc05cdfb
Pygments==2.1.3 --hash=sha256:485602129949b14247e8b124d28af4654dffbd076537c4a9c44a538a2c1755b7
docutils==0.12 --hash=sha256:dcebd4928112631626f4c4d0df59787c748404e66dda952110030ea883d3b8cd
alabaster==0.7.8 --hash=sha256:e86ae2eb1a0297454ce6bbe94414fbb272ed9f148b9ca5344dc213164e0d3d74
babel==2.3.4 --hash=sha256:3318ed2960240d61cbc6558858ee00c10eed77a6508c4d1ed8e6f7f48399c975
snowballstemmer==1.2.1 --hash=sha256:9f3bcd3c401c3e862ec0ebe6d2c069ebc012ce142cce209c098ccb5b09136e89
pytz==2016.6.1 --hash=sha256:7833bf559800232d3965b70e69642ebdadc76f7988f8d0a1440e072193ecd949
urlwait==0.3 \
    --hash=sha256:2b1fb4e5085344f91f1aaaa3c1809e44d100691b20c2431cb81af1824d82939c \
    --hash=sha256:54eb1cc8346f24fd906d379447b49c25a19af220cf7be35c8bac136ee8f09105
flake8==3.0.4 \
    --hash=sha256:603a3ae7c8030219fee084728ca02a8bbd3a51829cacf97b445172a46cb04662 \
    --hash=sha256:b4c210c998f07d6ff24325dd91fbc011f2c37bcd6bf576b188de01d8656e970d
mccabe==0.5.2 \
    --hash=sha256:91cc38b2c7636aaf1903e06d96ee960fb3dff9ca3afc595627c9a638f8e86d2b \
    --hash=sha256:3473f06c8b757bbb5cdf295099bf64032e5f7d6fe0ec2f97ee9b23cb0a435aff
pycodestyle==2.0.0 \
    --hash=sha256:2ce83f2046f5ab85c652ceceddfbde7a64a909900989b4b43e92b10b743d0ce5 \
    --hash=sha256:37f0420b14630b0eaaf452978f3a6ea4816d787c3e6dcbba6fb255030adae2e7
pyflakes==1.2.3 \
    --hash=sha256:e87bac26c62ea5b45067cc89e4a12f56e1483f1f2cda17e7c9b375b9fd2f40da \
    --hash=sha256:2e4a1b636d8809d8f0a69f341acf15b2e401a3221ede11be439911d23ce2139e
requests_mock==1.2.0 \
    --hash=sha256:8c89d8db3a5ddbce6121bdeb665891ca27bb37d18bbadd96c6c11fb472f07df1 \
    --hash=sha256:7e49e8b896d0dd72405bdae913cacc0593e6f6c505083fc2ff479a0e0c0262ba
pytest-wholenodeid==0.2 \
    --hash=sha256:6bc4edaf560c8b45ac4669bb8b99d8fbbb7f6e5af8dc7727200cd4d3ef4668c2 \
    --hash=sha256:bdcb4df88ca182a19752ef95150c0881deaafc2c593ef67c61515894e5309206
